requirements.txt
streamlit
pandas
numpy
openpyxl
python-pptx